Поправил БД и классы соответственно

+++ src/main/java/ru/biosecure/wicket/dao/PersonDAO.java
@@ -10,5 +10,5 @@ public interface PersonDAO {
 
     public List<Person> listPerson();
 
-    public void removePerson(Integer id);
+    public void removePerson(long id);
 }
