+++ src/main/java/ru/biosecure/wicket/service/PersonServiceImpl.java
@@ -27,7 +27,7 @@ public class PersonServiceImpl implements PersonService {
     }
 
     @Transactional
-    public void removePerson(Integer id) {
+    public void removePerson(long id) {
         personDAO.removePerson(id);
     }
 }
